+++ common/readme.ts
@@ -1,4 +1,5 @@
 import { Command } from "jsr:@cliffy/command";
+import { exists } from "jsr:@std/fs";
 import { join } from "jsr:@std/path";
 
 /**
@@ -7,12 +8,19 @@ import { join } from "jsr:@std/path";
 function logReadme(command: Command, { depth = 1 } = {}) {
   console.log(`${"#".repeat(depth)} ${command.getName()}`);
   console.log();
-  console.log("```" + command.getHelp({ colors: false }) + "```");
+  console.log("```" + trimLines(command.getHelp({ colors: false })) + "```");
   for (const subCommand of command.getCommands()) {
     console.log();
     logReadme(subCommand, { depth: depth + 1 });
   }
 }
 
-const cli = await import(join(Deno.cwd(), "cli.ts"));
-logReadme(cli.command);
+function trimLines(text: string) {
+  return text.split("\n").map((line) => line.replace(/\s+$/, "")).join("\n");
+}
+
+const entrypoint = join(Deno.cwd(), "cli.ts");
+if (await exists(entrypoint)) {
+  const cli = await import(entrypoint);
+  logReadme(cli.command);
+}
