+++ deno.json
@@ -11,7 +11,10 @@
       "command": "mkdir -p $HOME/.local/bin && cp dist/* $HOME/.local/bin",
       "dependencies": ["build"]
     },
-    "test": "deno test --doc */*.ts"
+    "lint": "deno fmt --check && deno lint",
+    "typecheck": "deno check **/*.ts",
+    "test": "deno test -RWE --unstable-kv --doc */*.ts",
+    "ok": "deno task lint && deno task typecheck && deno task test"
   },
   "imports": {
     "jsr:@cliffy/ansi/colors": "jsr:@cliffy/ansi@1.0.0-rc.7/colors",
