+++ duolingo/deno.json
@@ -1,7 +1,7 @@
 {
   "name": "@tugrulates/duolingo",
   "version": "0.1.0",
-  "exports": "./cli.ts",
+  "exports": "./client.ts",
   "tasks": {
     "compile": "deno compile -R=$HOME/.scripts -W=$HOME/.scripts -E=HOME -N=www.duolingo.com --unstable-kv --no-prompt -o=../dist/ cli.ts",
     "readme": "deno run -A --unstable-kv ../common/readme.ts > README.md"
